% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_krige_model_CV_haversine.R
\name{auto_krige_model_CV_haversine}
\alias{auto_krige_model_CV_haversine}
\title{Fitting multiple variogram models using haversine distance (default) or euclidean.}
\usage{
auto_krige_model_CV_haversine(
  df,
  distance = "haversine",
  best_model_criteria = "MSE",
  max.dist,
  model_classes = c("matern", "exponential", "gaussian", "spherical"),
  weights_types = c("npairs", "cressie", "equal")
)
}
\arguments{
\item{df}{A dataframe that has 'value' and two coordinates columns 'lat', 'long' for haversine or 'x' and 'y' for euclidean.}

\item{distance}{Type of distance to be choosen.}

\item{best_model_criteria}{The criteria to be taken for the choosing of the best model.}

\item{model_classes}{Classes of variogram models to be considered.}

\item{weight_types}{Weights considered in the fitting of the variogram models.}
}
\value{
A list containing all the results of the combinations of the model and a gstat object of the best model.
}
\description{
Fitting multiple variogram models using haversine distance (default) or euclidean.
}
