% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_krige.R
\name{cv_krige_haversine}
\alias{cv_krige_haversine}
\title{'Leave-one-out' Cross-Validation method to measure model performance.}
\usage{
cv_krige_haversine(df, varg, distance = "haversine")
}
\arguments{
\item{df}{A dataframe that has 'value' and two coordinates columns 'lat', 'long' for haversine or 'x' and 'y' for euclidean.}

\item{varg}{A variogram object from the gstat library.}

\item{distance}{Type of distance to be choosen.}
}
\value{
A tibble containing data from the Cross-Validation.
}
\description{
'Leave-one-out' Cross-Validation method to measure model performance.
}
