% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variog_non_euclidean.R
\name{variog_non_euclidean}
\alias{variog_non_euclidean}
\title{Modified function from the geoR package to calculate the sample variogram with any distance-matrix.}
\usage{
variog_non_euclidean(
  geodata,
  coords = geodata$coords,
  data = geodata$data,
  uvec = "default",
  breaks = "default",
  trend = "cte",
  lambda = 1,
  option = c("bin", "cloud", "smooth"),
  estimator.type = c("classical", "modulus"),
  nugget.tolerance,
  max.dist,
  pairs.min = 2,
  bin.cloud = FALSE,
  direction = "omnidirectional",
  tolerance = pi/8,
  unit.angle = c("radians", "degrees"),
  angles = FALSE,
  messages,
  ...
)
}
\description{
Modified function from the geoR package to calculate the sample variogram with any distance-matrix.
}
