% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{vectorizing_fun}
\alias{vectorizing_fun}
\title{Vectorizing a Matrix.}
\usage{
vectorizing_fun(a)
}
\arguments{
\item{a}{A matrix}
}
\value{
A vectorized matrix, a vector.
}
\description{
Vectorizing a Matrix.
}
